+++ src/main/java/me/heldplayer/mods/wecui/client/region/PointPoly.java
@@ -1,8 +1,8 @@
 
 package me.heldplayer.mods.wecui.client.region;
 
-import me.heldplayer.util.HeldCore.client.RenderHelper;
 import net.minecraft.util.AxisAlignedBB;
+import net.specialattack.forge.core.client.RenderHelper;
 
 import org.lwjgl.opengl.GL11;
 
@@ -19,7 +19,7 @@ public class PointPoly extends Point {
         if (!this.isValid()) {
             return;
         }
-        AxisAlignedBB aabb = AxisAlignedBB.getBoundingBox((double) this.coord.posX, (double) this.region.min, (double) this.coord.posZ, (double) this.coord.posX + 1.0D, (double) this.region.max + 1.0D, (double) this.coord.posZ + 1.0D).expand(0.02D, 0.02D, 0.02D).offset(-offsetX, -offsetY, -offsetZ);
+        AxisAlignedBB aabb = AxisAlignedBB.getBoundingBox(this.coord.posX, this.region.min, this.coord.posZ, this.coord.posX + 1.0D, this.region.max + 1.0D, this.coord.posZ + 1.0D).expand(0.02D, 0.02D, 0.02D).offset(-offsetX, -offsetY, -offsetZ);
 
         GL11.glColor4f(this.color.red, this.color.green, this.color.blue, opacity);
         RenderHelper.drawBox(aabb);
